% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mafft}
\alias{mafft}
\title{mafft}
\usage{
mafft(arglist = arglist_get(...))
}
\arguments{
\item{arglist}{Arguments as character vector passed to mafft}
}
\description{
Run mafft
}
\examples{
library(outsider)
mafft <- module_import('mafft', repo = 'dombennett/om..mafft')
mafft('--help')
}
