% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mafft}
\alias{mafft}
\title{mafft}
\usage{
mafft(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run mafft
}
\examples{
library(outsider)
mafft <- module_import('mafft', repo = 'dombennett/om..mafft..7.407')
mafft('--help')
}
