% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mafft}
\alias{mafft}
\title{mafft}
\usage{
mafft(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run mafft
}
\examples{
library(outsider)
raxml <- module_import('raxml', repo = 'dombennett/om..raxml..8.2.12')
raxml('-h')
}
